Revert "Use the virtual target for change events to avoid restoring controlled state on the real target (#10444)" (#10504)
This reverts the meaningful (src, non-test) part of commit 8116a3a5c840e45986ab37836fd90c0e907f3873 since we've reverted the commit it depended on in 7250fef4014ac953f7d5ccad58aaa21a6f52634b. I don't fully understand the negative implications of leaving this unreverted but it sounds like consensus is that it's safer to revert.

I left the new fixture and verified it works correctly in Chrome 62 Mac, as well as the jest tests passing.