Allow functions to be passed to custom element setters (#23042)
This is part of the new custom element features that were implemented
here:
https://github.com/facebook/react/commit/2ccb735d40289d6e41a699ec5c646fa86e786a7b

When a custom element has a setter for a property and passes the `in`
heuristic, the value passed to the react property should be assigned
directly to the custom element's property, regardless of the type of the
value. However, it was discovered that this isn't working with
functions. This patch makes it work with functions.

Fixes https://github.com/facebook/react/issues/23041