Revert "Hold host functions in var" (#26079)
Revert https://github.com/facebook/react/commit/7db76f13c. The commit
breaks old React Native where `nativeFabricUIManager` is undefined. I
need to add unit test for this to make sure it doesn't happen in the
future and create a mechanism to deal with undefined
`nativeFabricUIManager`.
This is to unblock React sync to React Native.