compiler: Scope-based reordering
Extends instruction reordering further to allow reordering of instructions in different scopes. The basic idea is that each instructions within a scope maintain their original order with respect to each other, but instructions for different scopes can be reordered relative to other scopes. This allows interleaved or nested scopes to be reordered to avoid the interleaving/nesting.

There are lot of things still very broken about this, but the one new fixture shows an example of this working.

Note that this approach works very well for interleaving. We'll probably need to add a bit more to handle nesting - in general it's better to move nested scopes first before their parent scope, but the current algorithm for reordering doesn't take this into account yet.

ghstack-source-id: 9dc8e62d31a946f2630545b2570eab47ac37f075
Pull Request resolved: https://github.com/facebook/react/pull/29803