[compiler] Use dependencies from source for useMemo scopes
Summary: This modified the behavior of the compiler when preserving source-level useMemos as reactive scopes to use the depencies from the source as well. This accounts for the possibility that the useMemo in the source is more general than is required by local code, but is necessary due to rules of react violations.

With this change, ideally the disableMemoization mode will behave exactly like the uncompiled code.

[ghstack-poisoned]