[compiler] InferReferenceEffects outputs a disjoint set of aliases
Test Plan:
For the purposes of a pass added later in this stack, the InferReferenceEffects now outputs a DisjointSet of identifier ids representing aliases. We add this to InferReferenceEffects because this pass already does its own alias analysis, so making it available to later passes is convenient.

This entails implementing copy() and equals() methods on DIsjointSets so that the InferenceState copy and merge methods can handle alias sets.

ghstack-source-id: 54b44cea8ca445856638bed9fc8cb79e43431b55
Pull Request resolved: https://github.com/facebook/react/pull/30974