Revert "[Re-land] Make prerendering always non-blocking (#31268)" (#31355)
This reverts commit 1e96cde339abc8663157ed472fa40fe2dc8cbc07.

It looked like the bug we found on the original land was related to
broken product code. But through landing #31268 we found additional bugs
internally. Since disabling the feature flag does not fix the bugs, we
have to revert again to unblock the sync. We can continue to debug with
our internal build.