[compiler] Allow sequence expressions in for loops, and prevent DCE of declarations when the variable is reassigned in a value block
Summary:
Two somewhat unrelated fixes that combine to allow loops like `for (i =1, j =1; ...)` to be handled. We introduce a new case for building reactive functions when a block prior to a goto doesnt generate a value; this is the case for the fallthrough of sequence blocks in for loop inits, since the value is not read by anything.

We also make it so that we don't DCE declarations of variables that are written to in value blocks. If we did DCE such declarations, we would then later convert the reassignment to a const declaration--but declarations in value blocks are banned.

[ghstack-poisoned]