[compiler] ReactiveIR: add Let/Assign/Branch/Join nodes
The implementation ensures writes are ordered, but is missing the logic to ensure that reads are ordered. The easy thing would be to order reads the same as writes, but we want to be able to order reads relative to each other — the real ordering constraint is that an instruction read the correct write. That will take a bit more work, for a follow-up.

[ghstack-poisoned]